transcription-service/requirements.txt
deepgram-sdk
flask
flask_socketio
python-dotenv